Add new modules.
