+++ redwood-treehouse-compose/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-treehouse-compose
+POM_NAME=Redwood Treehouse Compose
