+++ redwood-treehouse-configuration/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-treehouse-configuration
+POM_NAME=Redwood Treehouse Configuration
