+++ samples/emoji-search/presenter/build.gradle
@@ -10,6 +10,8 @@ kotlin {
     commonMain {
       dependencies {
         api projects.samples.emojiSearch.schema.compose
+        api projects.redwoodTreehouse
+        api projects.redwoodTreehouseLazylayoutApi
       }
     }
   }
