+++ settings.gradle
@@ -23,7 +23,9 @@ include ':redwood-tooling-codegen'
 include ':redwood-tooling-lint'
 include ':redwood-tooling-schema'
 include ':redwood-treehouse'
+include ':redwood-treehouse-compose'
 include ':redwood-treehouse-composeui'
+include ':redwood-treehouse-configuration'
 include ':redwood-treehouse-lazylayout-api'
 include ':redwood-treehouse-lazylayout-compose'
 include ':redwood-treehouse-lazylayout-composeui'
